\name{weatherfiles}
\alias{weatherfiles}
\title{All weather files}
\usage{
  weatherfiles()
}
\value{
  A vector with all available file names.
}
\description{
  Get all available weather file names
}
\examples{
head(weatherfiles)
length(weatherfiles)
}
\author{
  Laurent Gatto <lg390@cam.ac.uk>
}

